+++ packages/sandbox-e2b/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sandagent/sandbox-e2b
 
+## 0.9.27-beta.0
+
+### Patch Changes
+
+- @sandagent/manager@0.9.27-beta.0
+
 ## 0.9.26-beta.0
 
 ### Patch Changes